Changed iterator to be without index
darcs-hash:20081215105639-fafb9-a0f40e747b5569ef4b36ffbeb2d2dc7cf85b0d4b.gz

+++ testSpacing.cxx
@@ -51,8 +51,8 @@ int main(int, char * argv[])
   changer->SetInput(reader->GetOutput());
   ChangeType::SpacingType newspacing;
 
-  newspacing[0] = 1/sqrt(1);
-  newspacing[1] = 1/sqrt(0.5);
+  newspacing[0] = 1/sqrt((float)1);
+  newspacing[1] = 1/sqrt((float)0.5);
 
 
   changer->SetOutputSpacing(newspacing);
